Extract friends tab to its own component
