+++ angular-client/src/app/components/social-friends-tab/social-friends-tab.component.css
@@ -1,0 +1,49 @@
+table {
+  width: 100%;
+  background-color: #212529;
+}
+
+.mat-cell {
+  color: aliceblue;
+}
+
+.social-friends-tab {
+  padding-top: 10px;
+}
+
+/* get rid of padding left for the table */
+th.mat-header-cell:first-of-type,
+td.mat-cell:first-of-type,
+td.mat-footer-cell:first-of-type,
+td.mat-cell:last-of-type {
+  padding-left: 0px;
+  padding-right: 0px;
+}
+
+.right-justify {
+  text-align: right;
+}
+
+.mat-expansion-panel {
+  background-color: #212529;
+  color: aliceblue;
+}
+
+.mat-expansion-panel-header {
+  padding: 0px;
+  padding-right: 7px;
+}
+
+.mat-expansion-panel-header-description {
+  flex-grow: 0;
+  color: rgb(204, 33, 33);
+}
+
+::ng-deep .mat-expansion-panel-body {
+  padding: 0px !important;
+  padding-bottom: 7px !important;
+}
+
+::ng-deep .mat-expansion-indicator:after {
+  color: rgba(255, 255, 255, 0.5);
+}
