+++ angular-client/src/app/components/social-sidenav/social-sidenav.component.css
@@ -20,54 +20,8 @@
   display: none;
 } */
 
-table {
-  width: 100%;
-  background-color: #212529;
-}
-
-.mat-cell {
-  color: aliceblue;
-}
-
-/* get rid of padding left for the table */
-th.mat-header-cell:first-of-type,
-td.mat-cell:first-of-type,
-td.mat-footer-cell:first-of-type,
-td.mat-cell:last-of-type {
-  padding-left: 0px;
-  padding-right: 0px;
-}
-
-.right-justify {
-  text-align: right;
-}
-
 .profile-icon {
   padding-bottom: 5px;
-}
-
-.mat-expansion-panel {
-  background-color: #212529;
-  color: aliceblue;
-}
-
-.mat-expansion-panel-header {
-  padding: 0px;
-  padding-right: 7px;
-}
-
-.mat-expansion-panel-header-description {
-  flex-grow: 0;
-  color: rgb(204, 33, 33);
-}
-
-::ng-deep .mat-expansion-panel-body {
-  padding: 0px !important;
-  padding-bottom: 7px !important;
-}
-
-::ng-deep .mat-expansion-indicator:after {
-  color: rgba(255, 255, 255, 0.5);
 }
 
 .social-sidenav-tab-group ::ng-deep .mat-tab-label {
